Introducing NavLinks
* First, it was observed that the server-side routing was in place.
This led to 404 errors for all pages, since there isn't any route defined by the server.
Turning the bool value of historyApiFallback to true in webpack.config.js,
this was handled, and if the above 404 error was encountered, the index.html was served
instead, and the client side routers were checked, and served.
* A new 404 route was also introduced, for handling undefined routes.
* Switch was introduced for all the routers, so that 404 is displayed only when necessary.
* A new Header component has been introduced, to hold all the links that
must be rendered
* Links and NavLinks introduced, for better routing, and these have been put in the Header
component.