Create Mocks - moment()
In order to test if the state is updated correctly, we need mocks of
moment (since the time we test at will always be different from the
time defined in the state (the local state defined in the class
component)). This commit covers that, and also has introduced all tests
for ensuring that the above mentioned state is modified correctly.