Introduce Firebase support
Firebase support has been introduced. Firebase offers a realtime database service, which will be used to keep an account of all the expenses users add.
This is achieved by having an action dispatch a function instead of dispatching an object. This action is called by the AddExpense component.
The function that gets returned by the action, which is then handled by the reducer, first pushes the data into the database, and then proceeds to modify
store accordingly (used a good ol' fashioned action generator that returns an object)