Create new actions, setStartExpenses and startExpenses
Until the expenses are retrieved from the database, a message saying 'Loading!' is displayed.
setStartExpenses is a function, which will be dispatched by the store.
setStartExpenses fetches the data present in the db, parses it into an array, and dispatches the startExpenses action passing the parsed array of expenses as the only argument.
startExpenses will return the array of expenses that is returned, and thus, the store is updated, and once this is finally returned, the a promise by the setStartExpenses,
upon which the Dashboard is displayed.