Remove expenses on the DB using startRemoveExpenses
startRemoveExpenses is an action that gets dispatched and returns a function. It is similar in a lot of ways to startSetExpenses and startAddExpenses.
Check those commits for more details on the logic flow of this implementation.
startRemoveExpenses facilitates removal of the corresponding expense entry on the Firebase Database, when the user chooses to do so.
It goes without mentioning that the respective testcases have been modified as well; as always.